+++ esp-idf-svc/src/private/stubs.rs
@@ -1,6 +1,8 @@
 use core::ffi;
 
 // TODO: Figure out which library references this
+// picolibc already provides timegm, only define stub for newlib
+#[cfg(not(esp_idf_libc_picolibc))]
 #[no_mangle]
 pub extern "C" fn timegm(_: ffi::c_void) -> ffi::c_int {
     // Not supported but don't crash just in case