+++ esp-idf-svc/src/wifi.rs
@@ -1182,7 +1182,10 @@ impl<'d> WifiDriver<'d> {
         }
 
         esp!(unsafe { esp_wifi_wps_enable(&config.0 as *const _) })?;
+        #[cfg(not(esp_idf_version_at_least_6_0_0))]
         esp!(unsafe { esp_wifi_wps_start(0) })?;
+        #[cfg(esp_idf_version_at_least_6_0_0)]
+        esp!(unsafe { esp_wifi_wps_start() })?;
 
         self.status.lock().wps = None;